// Source/Microsoft.Teams.Apps.RewardAndRecognition/ClientApp/src/models/result.ts
/*
    <copyright file="result.ts" company="Microsoft">
    Copyright (c) Microsoft. All rights reserved.
    </copyright>
*/

export class ResultDetails {
    AwardId: string | undefined;
    AwardName: string | undefined;
    WinnerCount: number | undefined;
    NominationId: string | undefined;
    TeamId: string | undefined;
    NominatedToName: string | undefined;
    NominatedToObjectId: string | undefined;
    NominatedToPrincipalName: string | undefined;
    AwardLink: string | undefined;
    AwardCycle: string | undefined;
}